[refactor] Move matchUpdateCallback into PacketManager
